requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
joblib==1.1.0
scikit-learn==1.0.2
